list bets after betting
